+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
  - using Github Actions for testing and deployment
+ - `containment_to_tikz` method in `Perm` that returns an iterator of tikz pictures
+   displaying occurrences of classical patterns in self.
 
 ## 2.0.2 - 2020-08-06
 ### Fixed