+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
  - `containment_to_tikz` method in `Perm` that returns an iterator of tikz pictures
    displaying occurrences of classical patterns in self.
 
+### Deprecated
+ - Python 3.6 is no longer supported
+
 ## 2.0.2 - 2020-08-06
 ### Fixed
   - Include the type hints in the pypi release.